Convert to Typescript as user-defined type guards (#18)
Javascript has changed a lot in the 10 years since this library was first written,
and the needs for this sort of library have changed. Typescript has made writing
JS much nicer, and so this library will now provide a set of user-defined type
guards intended for use with data of non-guaranteed shape, such as API responses.

This modernizes the codebase to use typescript, add jest for testing, and make
use of Github Actions to automate testing workflows.